# src/main/resources/WEB-INF/openapi/swagger.yaml
swagger: "2.0"
info:
  description: |
    Swathi's 2023 onboarding ticket for Confluent
  version: "v1"
  title: "Swathi's Calculator"

tags:
  - name: Calculator API
    description: Has four operations and an audit endpoint

host: "localhost:8080/calculator"
schemes:
  - "http"
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
paths:
  /add:
    get:
      tags:
        - Add floats
      summary: |
        Adds floating point numbers
      description: Given multiple floating point numbers, it adds them.
      operationId: add
      parameters:
        - name: inputs
          in: query
          required: true
          description: Any floats to subtract
          explode: true
          type: array
          collectionFormat: multi
          items:
            type: float
      responses:
        '200':
          description: Authenticated
        '404':
          description: Invalid input.
  /subtract:
    get:
      tags:
        - Subtract floats
      summary: |
        Subtracts floating point numbers
      description: Given floating point numbers, it subtracts them.
      operationId: subtract
      parameters:
        - name: inputs
          in: query
          required: true
          description: Any floats to subtract
          explode: true
          type: array
          collectionFormat: multi
          items:
            type: float
      responses:
        '200':
          description: Authenticated
        '404':
          description: Invalid input.
  /multiply:
    get:
      tags:
        - Multiply floats
      summary: |
        Multiplies floating point numbers
      description: Given floating point numbers, it multiplies them.
      operationId: multiply
      parameters:
        - name: inputs
          in: query
          required: true
          description: Any floats to multiply
          explode: true
          type: array
          collectionFormat: multi
          items:
            type: float
      responses:
        '200':
          description: Authenticated
        '404':
          description: Invalid input.
  /divide:
    get:
      tags:
        - Divides floats
      summary: |
        Divides floating point numbers
      description: Given floating point numbers, it divides them.
      operationId: divide
      parameters:
        - name: inputs
          in: query
          required: true
          description: Any floats to divide
          explode: true
          type: array
          collectionFormat: multi
          items:
            type: float
      responses:
        '200':
          description: Authenticated
        '404':
          description: Invalid input.
  /audit:
    get:
      tags:
        - Returns a list of previous calculations
      summary: |
        Returns a list of previous calculations.
      description: Returns a list of previous calculations. Requires username/password
      operationId: audit
      security:
        - BasicAuth: [ admin ]
      responses:
        '200':
          description: Authenticated
        '404':
          description: Invalid input.
        '401':
          description: Unauthorized request.